Changed float functions to double counterparts

+++ AS3/Sample_TempObstacles.cpp
@@ -1203,9 +1203,9 @@ bool Sample_TempObstacles::handleBuild()
 	cfg.cs = m_cellSize;
 	cfg.ch = m_cellHeight;
 	cfg.walkableSlopeAngle = m_agentMaxSlope;
-	cfg.walkableHeight = (int)ceilf(m_agentHeight / cfg.ch);
-	cfg.walkableClimb = (int)floorf(m_agentMaxClimb / cfg.ch);
-	cfg.walkableRadius = (int)ceilf(m_agentRadius / cfg.cs);
+	cfg.walkableHeight = (int)ceil(m_agentHeight / cfg.ch);
+	cfg.walkableClimb = (int)floor(m_agentMaxClimb / cfg.ch);
+	cfg.walkableRadius = (int)ceil(m_agentRadius / cfg.cs);
 	cfg.maxEdgeLen = (int)(m_edgeMaxLen / m_cellSize);
 	cfg.maxSimplificationError = m_edgeMaxError;
 	cfg.minRegionArea = (int)rcSqr(m_regionMinSize);		// Note: area = size*size
