+++ RecastDemo/Source/ChunkyTriMesh.cpp
@@ -259,7 +259,7 @@ static bool checkOverlapSegment(const double p[2], const double q[2],
 	
 	for (int i = 0; i < 2; i++)
 	{
-		if (fabsf(d[i]) < EPSILON)
+        if (fabs(d[i]) < EPSILON)
 		{
 			// Ray is parallel to slab. No hit if origin not within slab
 			if (p[i] < bmin[i] || p[i] > bmax[i])
